Fix two scaling bugs surfaced by the generated-graph stress tests
- Hoist resolveImports out of Interpreter.init's per-import loop. It
  resolves all imports each call, so calling it once per import made
  init O(I^2 * M). Drops a 500-import module from ~10s to ~1s.
- Widen exposed-item node indices from u16 to u32 (ExposedItems through
  ModuleEnv, CIR external-lookup payloads, MIR external-def symbol IDs,
  and the Check/Lower/Monomorphize lookups that feed them). Modules with
  more than ~10k symbols no longer overflow the u16 cast and abort.
- Rephrase the coordinator stuck-detection: it's a no-progress timeout,
  not a deadlock proof. Lift the magic numbers into named constants and
  update the diagnostic preamble and panic string to match.

+++ src/canonicalize/Expression.zig
@@ -142,7 +142,7 @@ pub const Expr = union(enum) {
     /// ```
     e_lookup_external: struct {
         module_idx: CIR.Import.Idx,
-        target_node_idx: u16,
+        target_node_idx: u32,
         ident_idx: Ident.Idx,
         region: Region,
     },
@@ -282,7 +282,7 @@ pub const Expr = union(enum) {
     /// ```
     e_nominal_external: struct {
         module_idx: CIR.Import.Idx,
-        target_node_idx: u16,
+        target_node_idx: u32,
         backing_expr: Expr.Idx,
         backing_type: NominalBackingType,
     },
